+++ cpp/matrix.hpp
@@ -3,6 +3,6 @@
 namespace matrix {
     struct Matrix {
         int size;
-        long double** body;
+        double** body;
     };
 }
