+++ cpp/memory_manager.hpp
@@ -7,12 +7,12 @@ namespace memory {
     public:
         virtual ~MemoryManager() {}
         virtual void cleanup_matrix(const matrix::Matrix &matrix) = 0;
-        virtual void cleanup_matrix(long double** matrix) = 0;
+        virtual void cleanup_matrix(double** matrix) = 0;
     };
 
     class MemoryManagerImpl : public MemoryManager {
     public:
         virtual void cleanup_matrix(const matrix::Matrix &matrix) override;
-        virtual void cleanup_matrix(long double** matrix) override;
+        virtual void cleanup_matrix(double** matrix) override;
     };
 }
